requirements.txt
Flask==2.0.1
gunicorn==20.1.0
tensorflow==2.3.0
pandas
scikit-learn==0.24.2
numpy
keras==2.4.3
gcsfs
#google-python-cloud-debugger
google-cloud-aiplatform
oauth2client

